// src/service/file/schemas/recentTaskResultTaskType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 업무 종류. (LECTURE_PLAN:놀이계획, LECTURE_PLAN_REPORT:놀이보고서, STUDENT_RECORD:관찰기록, STORY_BOARD:스토리보드, STUDENT_CLASSIFICATION:아이분류, ACTIVITY_CLASSIFICATION: 활동분류, PHOTO_COMPOSITION: 아이합성, DATA_ENCRYPTION : 초상권보호SKETCH_CREATION : 도안 생성 PHOTO_ARRANGEMENT : 사진 정리,)
 */
export type RecentTaskResultTaskType = (typeof RecentTaskResultTaskType)[keyof typeof RecentTaskResultTaskType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RecentTaskResultTaskType = {
  LECTURE_PLAN: 'LECTURE_PLAN',
  LECTURE_PLAN_REPORT: 'LECTURE_PLAN_REPORT',
  STUDENT_RECORD: 'STUDENT_RECORD',
  STORY_BOARD: 'STORY_BOARD',
  STUDENT_CLASSIFICATION: 'STUDENT_CLASSIFICATION',
  ACTIVITY_CLASSIFICATION: 'ACTIVITY_CLASSIFICATION',
  PHOTO_COMPOSITION: 'PHOTO_COMPOSITION',
  PRIVATE_DATA_ENCRYPTION: 'PRIVATE_DATA_ENCRYPTION',
  SKETCH_CREATION: 'SKETCH_CREATION',
  PHOTO_ALBUM: 'PHOTO_ALBUM',
  AI_WRITING: 'AI_WRITING',
  STUDENT_AND_ACTIVITY_CLASSIFICATION: 'STUDENT_AND_ACTIVITY_CLASSIFICATION',
  PLAY_PHOTO: 'PLAY_PHOTO',
  REPLY: 'REPLY',
  ETC: 'ETC',
} as const;